Add style to header
The header, header logo a header links were given style.
The margin and padding of the body element were given a value of 0.
Decoration was removed from the 'a' elements.
And the 'ul' elements now hace no list-style, and are now flex.
The header links' text was modified in the index file.